23680961: tests into their repos
